move AttributeValueToString and AttributeMapToMap to pdata
