+++ model/pdata/common_test.go
@@ -1220,3 +1220,101 @@ func TestAnyValueArrayWithNilValues(t *testing.T) {
 	assert.EqualValues(t, AttributeValueTypeString, val.Type())
 	assert.EqualValues(t, "other_value", val.StringVal())
 }
+
+func TestAttributeValueToString(t *testing.T) {
+	tests := []struct {
+		name     string
+		input    AttributeValue
+		expected string
+	}{
+		{
+			name:     "string",
+			input:    NewAttributeValueString("string value"),
+			expected: "string value",
+		},
+		{
+			name:     "int64",
+			input:    NewAttributeValueInt(42),
+			expected: "42",
+		},
+		{
+			name:     "float64",
+			input:    NewAttributeValueDouble(1.61803399),
+			expected: "1.61803399",
+		},
+		{
+			name:     "boolean",
+			input:    NewAttributeValueBool(true),
+			expected: "true",
+		},
+		{
+			name:     "empty_map",
+			input:    NewAttributeValueMap(),
+			expected: "{}",
+		},
+		{
+			name:     "simple_map",
+			input:    simpleAttributeValueMap(),
+			expected: "{\"arrKey\":[\"strOne\",\"strTwo\"],\"boolKey\":false,\"floatKey\":18.6,\"intKey\":7,\"mapKey\":{\"keyOne\":\"valOne\",\"keyTwo\":\"valTwo\"},\"nullKey\":null,\"strKey\":\"strVal\"}",
+		},
+		{
+			name:     "empty_array",
+			input:    NewAttributeValueArray(),
+			expected: "[]",
+		},
+		{
+			name:     "simple_array",
+			input:    simpleAttributeValueArray(),
+			expected: "[\"strVal\",7,18.6,false,null]",
+		},
+		{
+			name:     "null",
+			input:    NewAttributeValueNull(),
+			expected: "",
+		},
+	}
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			actual := AttributeValueToString(test.input)
+			assert.Equal(t, test.expected, actual)
+		})
+	}
+}
+
+func simpleAttributeValueMap() AttributeValue {
+	ret := NewAttributeValueMap()
+	attrMap := ret.MapVal()
+	attrMap.UpsertString("strKey", "strVal")
+	attrMap.UpsertInt("intKey", 7)
+	attrMap.UpsertDouble("floatKey", 18.6)
+	attrMap.UpsertBool("boolKey", false)
+	attrMap.Upsert("nullKey", NewAttributeValueNull())
+	attrMap.Upsert("mapKey", constructTestAttributeSubmap())
+	attrMap.Upsert("arrKey", constructTestAttributeSubarray())
+	return ret
+}
+
+func simpleAttributeValueArray() AttributeValue {
+	ret := NewAttributeValueArray()
+	attrArr := ret.ArrayVal()
+	attrArr.AppendEmpty().SetStringVal("strVal")
+	attrArr.AppendEmpty().SetIntVal(7)
+	attrArr.AppendEmpty().SetDoubleVal(18.6)
+	attrArr.AppendEmpty().SetBoolVal(false)
+	attrArr.AppendEmpty()
+	return ret
+}
+
+func constructTestAttributeSubmap() AttributeValue {
+	value := NewAttributeValueMap()
+	value.MapVal().UpsertString("keyOne", "valOne")
+	value.MapVal().UpsertString("keyTwo", "valTwo")
+	return value
+}
+
+func constructTestAttributeSubarray() AttributeValue {
+	value := NewAttributeValueArray()
+	value.ArrayVal().AppendEmpty().SetStringVal("strOne")
+	value.ArrayVal().AppendEmpty().SetStringVal("strTwo")
+	return value
+}
