+++ translator/trace/protospan_translation.go
@@ -15,10 +15,6 @@
 package tracetranslator
 
 import (
-	"encoding/json"
-	"fmt"
-	"strconv"
-
 	"go.opentelemetry.io/collector/model/pdata"
 )
 
@@ -57,85 +53,6 @@ const (
 	OpenTracingSpanKindInternal    OpenTracingSpanKind = "internal"
 )
 
-// AttributeValueToString converts an OTLP AttributeValue object to its equivalent string representation
-func AttributeValueToString(attr pdata.AttributeValue) string {
-	switch attr.Type() {
-	case pdata.AttributeValueTypeNull:
-		return ""
-
-	case pdata.AttributeValueTypeString:
-		return attr.StringVal()
-
-	case pdata.AttributeValueTypeBool:
-		return strconv.FormatBool(attr.BoolVal())
-
-	case pdata.AttributeValueTypeDouble:
-		return strconv.FormatFloat(attr.DoubleVal(), 'f', -1, 64)
-
-	case pdata.AttributeValueTypeInt:
-		return strconv.FormatInt(attr.IntVal(), 10)
-
-	case pdata.AttributeValueTypeMap:
-		jsonStr, _ := json.Marshal(AttributeMapToMap(attr.MapVal()))
-		return string(jsonStr)
-
-	case pdata.AttributeValueTypeArray:
-		jsonStr, _ := json.Marshal(attributeArrayToSlice(attr.ArrayVal()))
-		return string(jsonStr)
-
-	default:
-		return fmt.Sprintf("<Unknown OpenTelemetry attribute value type %q>", attr.Type())
-	}
-}
-
-// AttributeMapToMap converts an OTLP AttributeMap to a standard go map
-func AttributeMapToMap(attrMap pdata.AttributeMap) map[string]interface{} {
-	rawMap := make(map[string]interface{})
-	attrMap.Range(func(k string, v pdata.AttributeValue) bool {
-		switch v.Type() {
-		case pdata.AttributeValueTypeString:
-			rawMap[k] = v.StringVal()
-		case pdata.AttributeValueTypeInt:
-			rawMap[k] = v.IntVal()
-		case pdata.AttributeValueTypeDouble:
-			rawMap[k] = v.DoubleVal()
-		case pdata.AttributeValueTypeBool:
-			rawMap[k] = v.BoolVal()
-		case pdata.AttributeValueTypeNull:
-			rawMap[k] = nil
-		case pdata.AttributeValueTypeMap:
-			rawMap[k] = AttributeMapToMap(v.MapVal())
-		case pdata.AttributeValueTypeArray:
-			rawMap[k] = attributeArrayToSlice(v.ArrayVal())
-		}
-		return true
-	})
-	return rawMap
-}
-
-// attributeArrayToSlice creates a slice out of a pdata.AnyValueArray.
-func attributeArrayToSlice(attrArray pdata.AnyValueArray) []interface{} {
-	rawSlice := make([]interface{}, 0, attrArray.Len())
-	for i := 0; i < attrArray.Len(); i++ {
-		v := attrArray.At(i)
-		switch v.Type() {
-		case pdata.AttributeValueTypeString:
-			rawSlice = append(rawSlice, v.StringVal())
-		case pdata.AttributeValueTypeInt:
-			rawSlice = append(rawSlice, v.IntVal())
-		case pdata.AttributeValueTypeDouble:
-			rawSlice = append(rawSlice, v.DoubleVal())
-		case pdata.AttributeValueTypeBool:
-			rawSlice = append(rawSlice, v.BoolVal())
-		case pdata.AttributeValueTypeNull:
-			rawSlice = append(rawSlice, nil)
-		default:
-			rawSlice = append(rawSlice, "<Invalid array value>")
-		}
-	}
-	return rawSlice
-}
-
 // StatusCodeFromHTTP takes an HTTP status code and return the appropriate OpenTelemetry status code
 // See: https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/semantic_conventions/http.md#status
 func StatusCodeFromHTTP(httpStatusCode int) pdata.StatusCode {
@@ -143,4 +60,16 @@ func StatusCodeFromHTTP(httpStatusCode int) pdata.StatusCode {
 		return pdata.StatusCodeUnset
 	}
 	return pdata.StatusCodeError
+}
+
+// AttributeValueToString converts an OTLP AttributeValue object to its equivalent string representation
+// Deprecated: use pdata.AttributeValueToString instead.
+func AttributeValueToString(attr pdata.AttributeValue) string {
+	return pdata.AttributeValueToString(attr)
+}
+
+// AttributeMapToMap converts an OTLP AttributeMap to a standard go map
+// Deprecated: use pdata.AttributeMapToMap instead.
+func AttributeMapToMap(attrMap pdata.AttributeMap) map[string]interface{} {
+	return pdata.AttributeMapToMap(attrMap)
 }
